First working version of ZMUTchange
Now changes the data of the scanned specimens. Uses drop down menus to input which fields to change, and text fields to input the changed value. Results are saved as csv, with a new csv file created every time the user selects different fields to change. (this prevents accidentally saving the values in the wrong column) Several consecutive changes can be made to the same specimen: the program forgets which QR codes it's read whenever the user selects different fields.